fixed bugs. Made charge_x and mass_x float

+++ kernel/particle/properties.hpp
@@ -5,8 +5,8 @@
 
 namespace particle {
   struct Properties {
-    unsigned int mass_x = 0; // in terms of unit mass
-    int charge_x = 0; // in terms of unit charge
+    float mass_x = 0; // in terms of unit mass
+    float charge_x = 0; // in terms of unit charge
     std::string name = "";
     std::string nickname = "";
   };
